feat: 글로벌 스타일 분리
