+++ src/App.js
@@ -1,68 +1,12 @@
 import React from "react";
-import reset from "styled-reset";
 import { Route } from "react-router-dom";
-import { ThemeProvider, createGlobalStyle } from "styled-components";
+import { ThemeProvider } from "styled-components";
 import theme from "styles/theme";
+import GlobalStyle from "styles/globalStyle";
 import StoryPage from "pages/StoryPage";
 import ParticipatePage from "pages/ParticipatePage";
 import LandingPage from "pages/LandingPage";
 import LinkPage from "pages/LinkPage";
-
-const GlobalStyle = createGlobalStyle`
-    ${reset}
-    *{
-      @font-face {
- font-family: 'NanumBarunGothic';
- font-style: normal;
- font-weight: 400;
- src: url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWeb.eot');
- src: url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWeb.eot?#iefix') format('embedded-opentype'), url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWeb.woff') format('woff'), url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWeb.ttf') format('truetype');
-}
-
-@font-face {
- font-family: 'NanumBarunGothic';
- font-style: normal;
- font-weight: 700;
- src: url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebBold.eot');
- src: url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebBold.eot?#iefix') format('embedded-opentype'), url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebBold.woff') format('woff'), url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebBold.ttf') format('truetype')
-}
-
-@font-face {
- font-family: 'NanumBarunGothic';
- font-style: normal;
- font-weight: 300;
- src: url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebLight.eot');
- src: url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebLight.eot?#iefix') format('embedded-opentype'), url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebLight.woff') format('woff'), url('//cdn.jsdelivr.net/font-nanumlight/1.0/NanumBarunGothicWebLight.ttf') format('truetype');
-}
-        box-sizing:border-box;
-        outline:none;
-      font-family: 'NanumSquare', sans-serif !important;
-
-    }
-    body{
-        font-size:14px;
-        height:100%;
-        background:#F0F0F0;
-        
-    }
-    a{
-        text-decoration:none;
-        color:inherit;
-        cursor: pointer;
-    }
-    ol, ul, li {
-        list-style: none;
-    }
-    img {
-        display: block;
-        width: 100%;
-        height: 100%;
-    }
-    input, button {
-        background-color: transparent;
-    }
- 
-`;
 
 const App = () => {
   return (
