+++ src/components/Logo.js
@@ -30,7 +30,10 @@ const Logo = React.memo(({ left, history }) => {
   const moveTo = useCallback(() => {
     if (location.pathname === "/linkPage") {
       history.goBack();
-    } else if (location.pathname === "/storyPage") {
+    } else if (
+      location.pathname === "/storyPage" ||
+      location.pathname === "/"
+    ) {
       history.push("/linkPage");
     }
   }, [history, location]);
