+++ src/containers/NotableContainer.js
@@ -1,0 +1,7 @@
+import React from "react";
+import NotableStoryList from "components/NotableStoryList";
+const NotableContainer = () => {
+  return <NotableStoryList></NotableStoryList>;
+};
+
+export default NotableContainer;
