+++ src/pages/LandingPage.js
@@ -2,6 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import LogoContainer from "containers/LogoContainer";
 import LandingContainer from "containers/LandingContainer";
+import NotableContainer from "containers/NotableContainer";
 import hamburgerBar from "images/hamburgerBar.png";
 const PageWrapper = styled.section`
   width: 100%;
@@ -14,6 +15,7 @@ const LandingPage = () => {
     <PageWrapper>
       <LogoContainer left={hamburgerBar}></LogoContainer>
       <LandingContainer />
+      <NotableContainer />
     </PageWrapper>
   );
 };
